+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/MediaScannerServiceUtils.java
@@ -234,6 +234,7 @@ public class MediaScannerServiceUtils {
     }
 
     public void updateSortOfArtist() {
+        mediaFileDao.clearArtistReadingOfDirectory();
         FixedIds merged = mergeSortOfArtist();
         FixedIds copied = copySortOfArtist();
         FixedIds compensated = compensateSortOfArtist();